+++ packages/camera/camera/lib/src/camera_controller.dart
@@ -488,13 +488,12 @@ class CameraController extends ValueNotifier<CameraValue> {
   /// Throws a [CameraException] if image streaming or video recording has
   /// already started.
   ///
-  /// The `startImageStream` method is only available on Android and iOS (other
-  /// platforms won't be supported in current setup).
+  /// The `startImageStream` method is only available on platforms that
+  /// report support for image streaming via [supportsImageStreaming].
   ///
   // TODO(bmparr): Add settings for resolution and fps.
   Future<void> startImageStream(onLatestImageAvailable onAvailable) async {
-    assert(defaultTargetPlatform == TargetPlatform.android ||
-        defaultTargetPlatform == TargetPlatform.iOS);
+    assert(supportsImageStreaming());
     _throwIfNotInitialized('startImageStream');
     if (value.isRecordingVideo) {
       throw CameraException(
@@ -526,11 +525,10 @@ class CameraController extends ValueNotifier<CameraValue> {
   /// Throws a [CameraException] if image streaming was not started or video
   /// recording was started.
   ///
-  /// The `stopImageStream` method is only available on Android and iOS (other
-  /// platforms won't be supported in current setup).
+  /// The `stopImageStream` method is only available on platforms that
+  /// report support for image streaming via [supportsImageStreaming].
   Future<void> stopImageStream() async {
-    assert(defaultTargetPlatform == TargetPlatform.android ||
-        defaultTargetPlatform == TargetPlatform.iOS);
+    assert(supportsImageStreaming());
     _throwIfNotInitialized('stopImageStream');
     if (!value.isStreamingImages) {
       throw CameraException(
@@ -885,6 +883,10 @@ class CameraController extends ValueNotifier<CameraValue> {
         .onCameraFocusing(_cameraId)
         .map((CameraFocusingEvent event) => event.currentlyFocusing);
   }
+
+  /// Check whether the camera platform supports image streaming.
+  bool supportsImageStreaming() =>
+      CameraPlatform.instance.supportsImageStreaming();
 
   /// Releases the resources of this camera.
   @override