+++ packages/path_provider/path_provider_android/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.2.16
+
+* Updates compileSdk 34 to flutter.compileSdkVersion.
+
+## 2.2.15
+
+* Removes unnecessary native code.
+
+## 2.2.14
+
+* Updates annotations lib to 1.9.1.
+
 ## 2.2.13
 
 * Updates annotations lib to 1.9.0.